chore: tidy dashboard css
